requirements.txt
absl-py==0.7.1
astor==0.7.1
astroid==2.2.5
attrs==19.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
blis==0.2.4
boto==2.49.0
boto3==1.9.155
botocore==1.12.155
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
comet-git-pure==0.19.11
comet-ml==1.0.52
configobj==5.0.6
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
defusedxml==0.5.0
docutils==0.14
en-core-web-sm==2.1.0
entrypoints==0.3
everett==1.0.2
gast==0.2.2
gensim==3.7.3
graphviz==0.10.1
grpcio==1.19.0
h5py==2.9.0
idna==2.8
ipykernel==5.1.0
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.17
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.3
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
murmurhash==1.0.2
nbconvert==5.4.1
nbformat==4.4.0
netifaces==0.10.9
nltk==3.4
notebook==5.7.8
numpy==1.16.2
nvidia-ml-py3==7.352.0
pandas==0.24.2
pandocfilters==1.4.2
pbr==5.1.3
pickleshare==0.7.5
Pillow==6.0.0
plac==0.9.6
plotly==3.8.1
preshed==2.0.1
prometheus-client==0.6.0
prompt-toolkit==1.0.15
protobuf==3.7.1
pydot==1.4.1
Pygments==2.3.1
pylint==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
retrying==1.3.3
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smart-open==1.8.3
soupsieve==1.9
spacy==2.1.3
srsly==0.0.5
tensorboard==1.12.2
tensorflow-estimator==1.13.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.0.4
torch==1.0.1
torchtext==0.3.1
torchvision==0.2.2.post3
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.24.1
wasabi==0.2.1
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.2
widgetsnbextension==3.4.2
win-unicode-console==0.5
wordcloud==1.5.0
wrapt==1.11.1
wurlitzer==1.0.2
